trying to update teal

+++ bin/x86_64-linux/bin/x86_64-linux/share/lua/5.1/luarocks/cmd/install.lua
@@ -2,6 +2,7 @@
 -- Installs binary rocks.
 local install = {}
 
+local dir = require("luarocks.dir")
 local path = require("luarocks.path")
 local repos = require("luarocks.repos")
 local fetch = require("luarocks.fetch")
@@ -13,7 +14,6 @@ local remove = require("luarocks.remove")
 local search = require("luarocks.search")
 local queries = require("luarocks.queries")
 local cfg = require("luarocks.core.cfg")
-local cmd = require("luarocks.cmd")
 
 function install.add_to_parser(parser)
    local cmd = parser:command("install", "Install a rock.", util.see_also())  -- luacheck: ignore 431
@@ -28,7 +28,8 @@ function install.add_to_parser(parser)
       "rock after building a new one. This behavior can be made permanent by "..
       "setting keep_other_versions=true in the configuration file.")
    cmd:flag("--force", "If --keep is not specified, force removal of "..
-      "previously installed versions if it would break dependencies.")
+      "previously installed versions if it would break dependencies. "..
+      "If rock is already installed, reinstall it anyway.")
    cmd:flag("--force-fast", "Like --force, but performs a forced removal "..
       "without reporting dependency issues.")
    cmd:flag("--only-deps --deps-only", "Install only the dependencies of the rock.")
@@ -84,6 +85,11 @@ function install.install_binary_rock(rock_file, opts)
       return nil, "Incompatible architecture "..arch, "arch"
    end
    if repos.is_installed(name, version) then
+      if not (opts.force or opts.force_fast) then
+         util.printout(name .. " " .. version .. " is already installed in " .. path.root_dir(cfg.root_dir))
+         util.printout("Use --force to reinstall.")
+         return name, version
+      end
       repos.delete_version(name, version, opts.deps_mode)
    end
 
@@ -119,7 +125,10 @@ function install.install_binary_rock(rock_file, opts)
    end
 
    if deps_mode ~= "none" then
-      ok, err, errcode = deps.fulfill_dependencies(rockspec, "dependencies", deps_mode, opts.verify, install_dir)
+      local deplock_dir = fs.exists(dir.path(".", "luarocks.lock"))
+                          and "."
+                          or install_dir
+      ok, err, errcode = deps.fulfill_dependencies(rockspec, "dependencies", deps_mode, opts.verify, deplock_dir)
       if err then return nil, err, errcode end
    end
 
@@ -220,9 +229,6 @@ end
 -- @return boolean or (nil, string, exitcode): True if installation was
 -- successful, nil and an error message otherwise. exitcode is optionally returned.
 function install.command(args)
-   local ok, err = fs.check_command_permissions(args)
-   if not ok then return nil, err, cmd.errorcodes.PERMISSIONDENIED end
-
    if args.rock:match("%.rockspec$") or args.rock:match("%.src%.rock$") then
       local build = require("luarocks.cmd.build")
       return build.command(args)
@@ -253,6 +259,13 @@ function install.command(args)
       args.rock = url
       return install.command(args)
    end
+end
+
+install.needs_lock = function(args)
+   if args.pack_binary_rock then
+      return false
+   end
+   return true
 end
 
 return install
